+++ src/DataStorage/Datastore.js
@@ -24,6 +24,8 @@ const defaultData = {
 
   UPGRADES: { BANK: 1 },
   DAILY: { LAST: "1999.00.00", STREAK: 0 },
+  PASSIVE_WORK_COLLECTION: {},
+
   SETTING_HIDDEN_FROM_LEADERBOARD: false,
 };